Relocate the default create location function
