+++ app/src/androidTest/java/ch/epfl/sdp/blindly/fake_module/FakeUserCacheModule.kt
@@ -2,6 +2,7 @@ package ch.epfl.sdp.blindly.fake_module
 
 import android.location.Location
 import ch.epfl.sdp.blindly.di.UserCacheModule
+import ch.epfl.sdp.blindly.location.AndroidLocationService.Companion.createLocationEPFL
 import ch.epfl.sdp.blindly.user.User
 import ch.epfl.sdp.blindly.user.UserCache
 import dagger.Module
@@ -10,9 +11,6 @@ import dagger.hilt.components.SingletonComponent
 import dagger.hilt.testing.TestInstallIn
 import org.mockito.Mockito
 import javax.inject.Singleton
-
-const val EPFL_LAT = 46.5
-const val EPFL_LONG = 6.5
 
 @Module
 @TestInstallIn(
@@ -43,13 +41,6 @@ open class FakeUserCacheModule {
                 .setDescription(description)
                 .setMatches(matches)
                 .build()
-
-        private fun createLocationEPFL(): Location {
-            val location = Location("")
-            location.latitude = EPFL_LAT
-            location.longitude = EPFL_LONG
-            return location
-        }
     }
 
     @Singleton
@@ -58,12 +49,5 @@ open class FakeUserCacheModule {
         val userCache = Mockito.mock(UserCache::class.java)
         Mockito.`when`(userCache.get(FakeUserHelperModule.TEST_UID)).thenReturn(fakeUser)
         return userCache
-    }
-
-    private fun createLocationEPFL(): Location {
-        val location = Location("")
-        location.latitude = 46.5
-        location.longitude = 6.5
-        return location
     }
 }
